move : changed package domain -> domain.role 패키지로 변경
